asim: move sim helper functions to helpers_test.go
This patch moves helper functions for scanning and parsing arguments from
`datadriven_simulation_test.go` to `helpers_test.go`.

Note that this commit does not change any existing functionality.

Epic: None

Release Note: None
