+++ pkg/kv/kvserver/asim/tests/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_test(
     name = "tests_test",
-    srcs = ["datadriven_simulation_test.go"],
+    srcs = [
+        "datadriven_simulation_test.go",
+        "helpers_test.go",
+    ],
     args = ["-test.timeout=295s"],
     data = glob(["testdata/**"]),
     embed = [":tests"],
